#20 fix unittests
